+++ Backend/.env
@@ -1,0 +1,4 @@
+DB_HOST=localhost
+DB_USER=root
+DB_PASSWORD=mysql
+DB_NAME=task
