+++ Backend/db/db.connect.ts
@@ -1,0 +1,21 @@
+import mysql from "mysql2";
+import dotenv from "dotenv";
+
+dotenv.config();
+
+export const db = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+});
+
+export const connectDB = () => {
+  db.connect((err) => {
+    if (err) {
+      console.error("Error connecting to MySQL:", err);
+      return;
+    }
+    console.log("Connected to MySQL");
+  });
+};
